[FIX] 약 정보 데이터 수정
